[SC-77] Pref: DB 테이블 info 하나로 통일하여 상속 entity 구성, 그에 따른 패키지 구조 변경
