+++ src/main/java/inu/codin/codin/domain/info/domain/lab/exception/LabNotFoundException.java
@@ -1,0 +1,8 @@
+package inu.codin.codin.domain.info.domain.lab.exception;
+
+public class LabNotFoundException extends RuntimeException{
+
+    public LabNotFoundException(String message){
+        super(message);
+    }
+}
