+++ src/main/java/inu/codin/codin/domain/info/domain/professor/exception/ProfessorNotFoundException.java
@@ -1,0 +1,7 @@
+package inu.codin.codin.domain.info.domain.professor.exception;
+
+public class ProfessorNotFoundException extends RuntimeException{
+    public ProfessorNotFoundException(String message){
+        super(message);
+    }
+}
