+++ src/main/java/inu/codin/codin/domain/info/entity/Info.java
@@ -1,0 +1,31 @@
+package inu.codin.codin.domain.info.entity;
+
+import inu.codin.codin.common.Department;
+import inu.codin.codin.domain.info.domain.professor.entity.Professor;
+import jakarta.validation.constraints.NotBlank;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+
+@Document(collection = "info")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public abstract class Info {
+
+    @Id @NotBlank
+    private String id;
+
+    @NotBlank
+    private Department department;
+
+    @NotBlank
+    private InfoType infoType;
+
+    public Info(String id, Department department, InfoType infoType) {
+        this.id = id;
+        this.department = department;
+        this.infoType = infoType;
+    }
+}
