--- src/main/java/inu/codin/codin/domain/info/lab/exception/LabNotFoundException.java
@@ -1,8 +1,0 @@
-package inu.codin.codin.domain.info.lab.exception;
-
-public class LabNotFoundException extends RuntimeException{
-
-    public LabNotFoundException(String message){
-        super(message);
-    }
-}
