--- src/main/java/inu/codin/codin/domain/info/lab/repository/LabRepository.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.info.lab.repository;
-
-import inu.codin.codin.domain.info.lab.entity.Lab;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface LabRepository extends MongoRepository<Lab, String> {
-}
