--- src/main/java/inu/codin/codin/domain/info/office/dto/OfficeMember.java
@@ -1,29 +1,0 @@
-package inu.codin.codin.domain.info.office.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-
-/*
-    학과 사무실 직원 정보
- */
-@Getter
-public class OfficeMember {
-    @NotBlank
-    @Schema(description = "성명", example = "홍길동")
-    private String name;
-    @NotBlank
-    @Schema(description = "직위", example = "조교")
-    private String position;
-
-    @Schema(description = "담당 업무", example = "학과사무실 업무")
-    private String role;
-
-    @NotBlank
-    @Schema(description = "연락처", example = "032-123-2345")
-    private String number;
-
-    @NotBlank
-    @Schema(description = "이메일", example = "test@inu.ac.kr")
-    private String email;
-}
