--- src/main/java/inu/codin/codin/domain/info/office/repository/OfficeRepository.java
@@ -1,13 +1,0 @@
-package inu.codin.codin.domain.info.office.repository;
-
-import inu.codin.codin.domain.info.office.entity.Office;
-import inu.codin.codin.common.Department;
-import jakarta.validation.constraints.NotBlank;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import java.util.List;
-
-public interface OfficeRepository extends MongoRepository<Office, String> {
-
-    List<Office> findAllByDepartment(@NotBlank Department department);
-}
