--- src/main/java/inu/codin/codin/domain/info/professor/exception/ProfessorNotFoundException.java
@@ -1,7 +1,0 @@
-package inu.codin.codin.domain.info.professor.exception;
-
-public class ProfessorNotFoundException extends RuntimeException{
-    public ProfessorNotFoundException(String message){
-        super(message);
-    }
-}
