--- src/main/java/inu/codin/codin/domain/info/professor/service/ProfessorService.java
@@ -1,30 +1,0 @@
-package inu.codin.codin.domain.info.professor.service;
-
-import inu.codin.codin.domain.info.professor.dto.ProfessorListResDTO;
-import inu.codin.codin.domain.info.professor.dto.ProfessorThumbnailResDTO;
-import inu.codin.codin.domain.info.professor.entity.Professor;
-import inu.codin.codin.domain.info.professor.exception.ProfessorNotFoundException;
-import inu.codin.codin.domain.info.professor.repository.ProfessorRepository;
-import inu.codin.codin.common.Department;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class ProfessorService {
-
-    private final ProfessorRepository professorRepository;
-
-    public List<ProfessorListResDTO> getProfessorByDepartment(Department department){
-        List<Professor> professors = professorRepository.findAllByDepartment(department);
-        return professors.stream().map(ProfessorListResDTO::of).toList();
-    }
-
-    public ProfessorThumbnailResDTO getProfessorThumbnail(String id) {
-        Professor professor = professorRepository.findById(id)
-                .orElseThrow(() -> new ProfessorNotFoundException("교수 정보를 찾을 수 없습니다."));
-        return ProfessorThumbnailResDTO.of(professor);
-    }
-}
